[DEL] 마이페이지에서 필요없는파일제거(#85)
