+++ Plu-iOS/Plu-iOS/Scenes/MyPage/Enum/MyPageSection.swift
@@ -7,21 +7,21 @@
 
 import Foundation
 
-enum MyPageSection {
-    case alarm(MyPageAlarmData)
-    case info(MyPageInfomaitonType)
-    case appVersion(MypageAppVersionData)
-    case exit(MyPageUserExitType)
-}
-
-extension MyPageSection {
-    static func makeMypageData(_ alarmAccept: Bool, _ appVersion: String?) -> [[Self]] {
-        return [[.alarm(.init(.alarm, acceptAlarm: alarmAccept))],
-                [.info(.faq),
-                 .info(.openSource),
-                 .info(.privacy)],
-                [.appVersion(.init(.appVersion, appVersion: appVersion))],
-                [.exit(.logout),
-                 .exit(.resign)]]
-    }
-}
+//enum MyPageSection {
+//    case alarm(MyPageAlarmData)
+//    case info(MyPageInfomaitonType)
+//    case appVersion(MypageAppVersionData)
+//    case exit(MyPageUserExitType)
+//}
+//
+//extension MyPageSection {
+//    static func makeMypageData(_ alarmAccept: Bool, _ appVersion: String?) -> [[Self]] {
+//        return [[.alarm(.init(.alarm, acceptAlarm: alarmAccept))],
+//                [.info(.faq),
+//                 .info(.openSource),
+//                 .info(.privacy)],
+//                [.appVersion(.init(.appVersion, appVersion: appVersion))],
+//                [.exit(.logout),
+//                 .exit(.resign)]]
+//    }
+//}
