--- Plu-iOS/Plu-iOS/Scenes/MyPage/MyPageInerface/MyPageCell.swift
@@ -1,12 +1,0 @@
-//
-//  MyPageCell.swift
-//  Plu-iOS
-//
-//  Created by uiskim on 2023/12/06.
-//
-
-import Foundation
-
-protocol MyPageCell {
-    var title: String { get }
-}
