--- Plu-iOS/Plu-iOS/Scenes/MyPage/MyPageInerface/MyPagePresentable.swift
@@ -1,13 +1,0 @@
-//
-//  MyPagePresentable.swift
-//  Plu-iOS
-//
-//  Created by uiskim on 2023/12/24.
-//
-
-import Foundation
-
-protocol MyPagePresentable {
-    var tableData: [[MyPageSection]] { get set }
-}
-
